CAT-2674 BUG Solved undo redo button for seekbars
Deleted the formulaEditorEditText.endEdit() and some other parts in setInputFormula function in FormulaEditorFragment.java because they were reduntant. In addition deleted the currentMenu in onPrepareOptionsMenu because it is not used. Changed in FormulaEditorEdittext the history object to an array of 3 objects. For each color e.g in seekbar a history. Introduced a new variable named laststate to determine which history of which seekbar was used recently. Made a OnClicklistener for the Colourtexts to switch between the seekbars.